+++ .vscode/settings.json
@@ -1,12 +1,19 @@
 {
-  "python.formatting.provider": "black",
   "[python]": {
+    "editor.defaultFormatter": "charliermarsh.ruff",
     "editor.formatOnSave": true,
     "editor.codeActionsOnSave": {
-      "source.showNotification": true,
-      "source.fixAll": true,
-      "source.organizeImports": true
-    }
+      "source.showNotification": "explicit",
+      "source.fixAll": "explicit",
+      "source.organizeImports": "explicit"
+    },
   },
-  "esbonio.sphinx.confDir": "docs/source"
+  "python.analysis.typeCheckingMode": "basic",
+  "python.testing.pytestArgs": [
+    "tests"
+  ],
+  "python.testing.unittestEnabled": false,
+  "python.testing.pytestEnabled": true
+  // },
+  // "esbonio.sphinx.confDir": "docs/source"
 }
