+++ README.md
@@ -1,5 +1,6 @@
 # hotpepper-gourmet
 
+![workflow badge](https://github.com/paperlefthand/hotpepper-gourmet/actions/workflows/pytest.yml/badge.svg)
 ![workflow badge](https://github.com/paperlefthand/hotpepper-gourmet/actions/workflows/deploy.yml/badge.svg)
 
 ## About
@@ -12,14 +13,16 @@
 
 ホットペッパーグルメAPIに登録し, token(keyid)を取得
 
-### サンプル
+### サンプルコード
 
 ``` python
-import pygourmet
-
-api = pygourmet.Api(keyid=YOUR_KEYID)
-results = api.search(lat=35.170915, lng=136.8793482, radius=400)
-print(results)
+>>> import pygourmet
+>>> api = pygourmet.Api(keyid=YOUR_KEYID)
+>>> results = api.search(lat=35.170915, lng=136.8793482, radius=400, count=3)
+>>> len(reaults)
+3
+>>> results[0]["name"]
+>>> 'shop name'
 ```
 
 ___
