+++ docs/source/resources/exceptions.rst
