+++ docs/source/resources/modules.rst
