+++ pygourmet/params.py
@@ -1,0 +1,6 @@
+class Params:
+    def __init__(self, key=None, format=None, count=None, start=None):
+        self.key = key
+        self.format = format
+        self.count = count
+        self.start = start
