+++ pyproject.toml
@@ -1,19 +1,20 @@
 [tool.poetry]
 name = "hotpepper-gourmet"
-version = "0.1.0"
+version = "0.0.1"
 description = "A simple Python wrapper for Hotpepper API"
 authors = ["paperlefthand <being.paperlefthand@gmail.com>"]
 license = "MIT"
 readme = "README.md"
 repository = "https://github.com/paperlefthand/hotpepper-gourmet"
-keywords = []
-classifiers = [
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Topic :: Software Development :: Libraries :: Python Modules",
-    "Intended Audience :: Developers",
-]
+# keywords = []
+# classifiers = [
+#     "Programming Language :: Python :: Implementation :: CPython",
+#     "Programming Language :: Python :: 3.10",
+#     "Programming Language :: Python :: 3.11",
+#     "Programming Language :: Python :: 3.12",
+#     "Topic :: Software Development :: Libraries :: Python Modules",
+#     "Intended Audience :: Developers",
+# ]
 packages = [
     { include = "pygourmet"},
 ]
@@ -23,14 +24,16 @@ python = "^3.10"
 requests = "^2.28"
 
 [tool.poetry.group.dev.dependencies]
-pytest-testinfra = "^7.0"
-responses = "^0.22"
-black = "^23.1.0"
-ruff = "^0.0.259"
-sphinx-rtd-theme = "^1.2.0"
-sphinx-fontawesome = "^0.0.6"
-myst-parser = "^1.0.0"
-sphinx-autobuild = "^2021.3.14"
+pytest = "^8.0"
+# pytest-testinfra = "^7.0"
+# responses = "^0.22"
+# black = "^23.1.0"
+ruff = "^0.2.0"
+# sphinx-rtd-theme = "^1.2.0"
+# sphinx-fontawesome = "^0.0.6"
+# myst-parser = "^1.0.0"
+# sphinx-autobuild = "^2021.3.14"
+requests-mock = "^1.11.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
